Refactor event validators into types
Before we had one validator 'businessEventvalidator' but the
validators can be more fine grained. So refactored the validator
MandatoryEventValidator, ReminderEventValidator and
ReferncePeriodEventValidator.
