+++ src/main/java/uk/gov/ons/ctp/response/collection/exercise/CollectionExerciseApplication.java
@@ -31,8 +31,6 @@ import uk.gov.ons.ctp.response.collection.exercise.representation.CollectionExer
 import uk.gov.ons.ctp.response.collection.exercise.representation.CollectionExerciseDTO.CollectionExerciseState;
 import uk.gov.ons.ctp.response.collection.exercise.representation.SampleUnitGroupDTO.SampleUnitGroupEvent;
 import uk.gov.ons.ctp.response.collection.exercise.representation.SampleUnitGroupDTO.SampleUnitGroupState;
-import uk.gov.ons.ctp.response.collection.exercise.service.EventValidator;
-import uk.gov.ons.ctp.response.collection.exercise.service.impl.BusinessEventValidator;
 import uk.gov.ons.ctp.response.collection.exercise.state.CollectionExerciseStateTransitionManagerFactory;
 
 /**
@@ -214,11 +212,6 @@ public class CollectionExerciseApplication implements CommandLineRunner {
     CustomObjectMapper mapper = new CustomObjectMapper();
 
     return mapper;
-  }
-
-  @Bean
-  public EventValidator eventValidator() {
-    return new BusinessEventValidator();
   }
 
   /**
